ana/requirements.txt
Django
djangorestframework
django-toolbelt
pytest
pytest-django
pytest-pythonpath
requests
jsonfield
